deprecate onboarding-enabler zutil, move to apiml-utility
